Extract RecurrenceSimulationMixin to de-duplicate simulation code
GeneralizedRenewal, GeneralizedOneRenewal and ParametricRecurrenceModel each
carried their own copies of the RNG helpers (initialize_simulation,
clear_simulation, get_uniform_random_number) and near-identical
count_terminated_simulation / time_terminated_simulation loops, which had also
drifted apart (events vs events+1 loop length, < vs <= count mask).

Introduce surpyval/recurrent/simulation.py with RecurrenceSimulationMixin that
owns the RNG helpers and both simulation loops. Each class now supplies only
its per-event sampling logic via a _new_sequence_sampler closure (G1 scale
factor, Kijima virtual age, or conditional-intensity inversion), and
ParametricRecurrenceModel overrides _postprocess_simulated_model for the
CoxLewis offset.

This also unifies the count-terminated convention (events+1 events per
sequence, mask mcf_hat < events) across all three. The only resulting numeric
change is GeneralizedRenewal's count output, whose docstring example is
regenerated; the Kaminskiy/Krivtsov G1 result and all other simulations are
bit-for-bit unchanged. Full suite: 528 passed.
