+++ surpyval/recurrent/parametric/parametric_recurrence.py
@@ -1,13 +1,10 @@
-import warnings
-
 import numpy as np
 from matplotlib import pyplot as plt
-from scipy.stats import uniform
 
-from surpyval.recurrent.nonparametric import NonParametricCounting
+from surpyval.recurrent.simulation import RecurrenceSimulationMixin
 
 
-class ParametricRecurrenceModel:
+class ParametricRecurrenceModel(RecurrenceSimulationMixin):
     """
     A class for holding the parameters, data, and usefult methods for a
     fitted parametric recurrence model. This is the result of the ``fit`` calls
@@ -40,161 +37,21 @@ class ParametricRecurrenceModel:
             + param_string
         )
 
-    def initialize_simulation(self):
-        self.us = uniform.rvs(size=100_000).tolist()
+    def _new_sequence_sampler(self):
+        x_prev = 0.0
 
-    def clear_simulation(self):
-        del self.us
+        def sample(ui):
+            nonlocal x_prev
+            u_adj = ui * np.exp(-self.cif(x_prev))
+            xi = self.inv_cif(-np.log(u_adj)) - x_prev
+            x_prev += xi
+            return xi
 
-    def get_uniform_random_number(self):
-        try:
-            return self.us.pop()
-        except IndexError:
-            self.initialize_simulation()
-            return self.us.pop()
+        return sample
 
-    def count_terminated_simulation(self, events, items=1):
-        """
-        Simulate count-terminated recurrence data based on the fitted model.
-
-        Parameters
-        ----------
-
-        events: int
-            Number of events to simulate.
-        items: int, optional
-            Number of items (or sequences) to simulate. Default is 1.
-
-        Returns
-        -------
-
-        NonParametricCounting
-            An NonParametricCounting model built from the simulated data.
-        """
-        self.initialize_simulation()
-
-        xicn = {"x": [], "i": [], "c": [], "n": []}
-
-        for i in range(0, items):
-            running = 0
-            j = 0
-            x_prev = 0
-            for j in range(0, events + 1):
-                ui = self.get_uniform_random_number()
-                u_adj = ui * np.exp(-self.cif(x_prev))
-                xi = self.inv_cif(-np.log(u_adj)) - x_prev
-                running += xi
-                x_prev = running
-                xicn["i"].append(i + 1)
-                xicn["n"].append(1)
-                xicn["x"].append(running)
-                xicn["c"].append(0)
-
-        self.clear_simulation()
-
-        model = NonParametricCounting.fit(**xicn)
-
+    def _postprocess_simulated_model(self, model):
         if self.dist.name == "CoxLewis":
             model.mcf_hat += np.exp(self.params[0])
-
-        mask = model.mcf_hat <= events
-        model.x = model.x[mask]
-        model.mcf_hat = model.mcf_hat[mask]
-        model.var = None
-        return model
-
-    def time_terminated_simulation(
-        self, T, items=1, tol=1e-8, max_events=10_000
-    ):
-        """
-        Simulate time-terminated recurrence data based on the fitted model.
-
-        Parameters
-        ----------
-
-        T: float
-            Time termination value.
-        items: int, optional
-            Number of items (or sequences) to simulate. Default is 1.
-        tol: float, optional
-            Interarrival times below this value end the sequence early; a tiny
-            increment indicates the cumulative time has stalled below T (a
-            possible asymptote). Default is 1e-8.
-        max_events: int, optional
-            Hard cap on the number of events simulated per sequence. This is
-            the backstop that guarantees termination for sequences whose
-            cumulative time cannot reach T. Default is 10000.
-
-        Returns
-        -------
-
-        NonParametricCounting
-            An NonParametricCounting model built from the simulated data.
-
-        Warnings
-        --------
-
-        A sequence is terminated early and right-censored at its last event if
-        an interarrival time falls below ``tol`` or it reaches ``max_events``
-        before T. A warning is raised in either case.
-        """
-        self.initialize_simulation()
-        stalled = False
-        hit_max_events = False
-
-        xicn = {"x": [], "i": [], "c": [], "n": []}
-
-        for i in range(0, items):
-            running = 0
-            n_events = 0
-            x_prev = 0
-            while True:
-                ui = self.get_uniform_random_number()
-                u_adj = ui * np.exp(-self.cif(x_prev))
-                xi = self.inv_cif(-np.log(u_adj)) - x_prev
-                running += xi
-                n_events += 1
-                x_prev = running
-                xicn["i"].append(i + 1)
-                xicn["n"].append(1)
-                if running > T:
-                    xicn["x"].append(T)
-                    xicn["c"].append(1)
-                    break
-                elif xi < tol:
-                    stalled = True
-                    xicn["x"].append(running)
-                    xicn["c"].append(0)
-                    break
-                elif n_events >= max_events:
-                    hit_max_events = True
-                    xicn["x"].append(running)
-                    xicn["c"].append(0)
-                    break
-                else:
-                    xicn["x"].append(running)
-                    xicn["c"].append(0)
-
-        self.clear_simulation()
-
-        if stalled:
-            warnings.warn(
-                "Some sequences produced a near-zero interarrival time "
-                "(< tol) before reaching T, indicating a possible asymptote; "
-                "they were terminated early at their last event."
-            )
-        if hit_max_events:
-            warnings.warn(
-                "Some sequences reached max_events ({}) before T; increase "
-                "max_events or check the model parameters.".format(max_events)
-            )
-
-        model = NonParametricCounting.fit(**xicn)
-
-        if self.dist.name == "CoxLewis":
-            model.mcf_hat += np.exp(self.params[0])
-        model.var = None
-
         return model
 
     def cif(self, x):
