+++ surpyval/recurrent/renewal/generalized_one_renewal.py
@@ -1,18 +1,15 @@
-import warnings
-
 import numpy as np
 from scipy.optimize import minimize
-from scipy.stats import uniform
 
 from surpyval import Weibull
-from surpyval.recurrent.nonparametric import NonParametricCounting
+from surpyval.recurrent.simulation import RecurrenceSimulationMixin
 from surpyval.utils.recurrent_utils import (
     handle_xicn,
     validate_renewal_censoring,
 )
 
 
-class GeneralizedOneRenewal:
+class GeneralizedOneRenewal(RecurrenceSimulationMixin):
     """
     A class to handle the G1 renewal process of Kaminskiy and Krivtsov, in
     which the jth interarrival time is the underlying lifetime distribution
@@ -89,160 +86,21 @@ class GeneralizedOneRenewal:
 
         return out
 
-    def initialize_simulation(self):
-        self.us = uniform.rvs(size=100_000).tolist()
-
-    def clear_simulation(self):
-        del self.us
-
-    def get_uniform_random_number(self):
-        try:
-            return self.us.pop()
-        except IndexError:
-            self.initialize_simulation()
-            return self.us.pop()
-
-    def count_terminated_simulation(self, events, items=1):
-        """
-        Simulate count-terminated recurrence data based on the fitted model.
-
-        Parameters
-        ----------
-
-        events: int
-            Number of events to simulate.
-        items: int, optional
-            Number of items (or sequences) to simulate. Default is 1.
-
-        Returns
-        -------
-
-        NonParametricCounting
-            An NonParametricCounting model built from the simulated data.
-        """
+    def _new_sequence_sampler(self):
         base_params = self.model.params
         q = self.q
-        self.initialize_simulation()
+        j = 0
 
-        xicn = {"x": [], "i": [], "c": [], "n": []}
+        def sample(ui):
+            nonlocal j
+            # The jth interarrival is the base lifetime scaled by (1 + q) ** j,
+            # so its quantiles are the base quantiles multiplied by the same
+            # factor.
+            cj = (1.0 + q) ** j
+            j += 1
+            return cj * self.model.dist.qf(ui, *base_params)
 
-        for i in range(0, items):
-            running = 0
-            for j in range(0, events + 1):
-                ui = self.get_uniform_random_number()
-                # The jth interarrival is the base lifetime scaled by
-                # (1 + q) ** j, so its quantiles are the base quantiles
-                # multiplied by the same factor.
-                cj = (1.0 + q) ** j
-                xi = cj * self.model.dist.qf(ui, *base_params)
-                running += xi
-                xicn["x"].append(running)
-                xicn["i"].append(i + 1)
-                xicn["c"].append(0)
-                xicn["n"].append(1)
-
-        self.clear_simulation()
-
-        model = NonParametricCounting.fit(**xicn)
-        mask = model.mcf_hat < events
-        model.x = model.x[mask]
-        model.mcf_hat = model.mcf_hat[mask]
-        model.var = None
-        return model
-
-    def time_terminated_simulation(
-        self, T, items=1, tol=1e-8, max_events=10_000
-    ):
-        """
-        Simulate time-terminated recurrence data based on the fitted model.
-
-        Parameters
-        ----------
-
-        T: float
-            Time termination value.
-        items: int, optional
-            Number of items (or sequences) to simulate. Default is 1.
-        tol: float, optional
-            Interarrival times below this value end the sequence early; a tiny
-            increment indicates the cumulative time has stalled below T (a
-            possible asymptote). Default is 1e-8.
-        max_events: int, optional
-            Hard cap on the number of events simulated per sequence. This is
-            the backstop that guarantees termination for sequences whose
-            cumulative time cannot reach T. Default is 10000.
-
-        Returns
-        -------
-
-        NonParametricCounting
-            An NonParametricCounting model built from the simulated data.
-
-        Warnings
-        --------
-
-        A sequence is terminated early and right-censored at its last event if
-        an interarrival time falls below ``tol`` or it reaches ``max_events``
-        before T. A warning is raised in either case.
-        """
-        base_params = self.model.params
-        q = self.q
-        self.initialize_simulation()
-        stalled = False
-        hit_max_events = False
-
-        xicn = {"x": [], "i": [], "c": [], "n": []}
-
-        for i in range(0, items):
-            running = 0
-            n_events = 0
-            while True:
-                ui = self.get_uniform_random_number()
-                # The jth interarrival is the base lifetime scaled by
-                # (1 + q) ** j, so its quantiles are the base quantiles
-                # multiplied by the same factor.
-                cj = (1.0 + q) ** n_events
-                xi = cj * self.model.dist.qf(ui, *base_params)
-                running += xi
-                n_events += 1
-                xicn["i"].append(i + 1)
-                xicn["n"].append(1)
-                if running > T:
-                    xicn["x"].append(T)
-                    xicn["c"].append(1)
-                    break
-                elif xi < tol:
-                    stalled = True
-                    xicn["x"].append(running)
-                    xicn["c"].append(0)
-                    break
-                elif n_events >= max_events:
-                    hit_max_events = True
-                    xicn["x"].append(running)
-                    xicn["c"].append(0)
-                    break
-                else:
-                    xicn["x"].append(running)
-                    xicn["c"].append(0)
-
-        self.clear_simulation()
-
-        if stalled:
-            warnings.warn(
-                "Some sequences produced a near-zero interarrival time "
-                "(< tol) before reaching T, indicating a possible asymptote; "
-                "they were terminated early at their last event."
-            )
-        if hit_max_events:
-            warnings.warn(
-                "Some sequences reached max_events ({}) before T; increase "
-                "max_events or check the model parameters.".format(max_events)
-            )
-
-        model = NonParametricCounting.fit(**xicn)
-        model.var = None
-
-        return model
+        return sample
 
     @classmethod
     def create_negll_func(cls, x, i, c, n, dist):
